requirements.txt
flask
rpyc
uwsgi
flask-restful
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
pysensors